Refactor scraper code into a seperate binary and move common code to libraries
Refactor to seperate out the db layer and the utils layer which could be used in other binaries. Renamed uschess binary to scraper binary to accurately reflect its work.
