--- utils.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"log"
-)
-
-func checkErr(err error) {
-	if err != nil {
-		log.Fatal(err)
-	}
-}
