+++ utils/utils.go
@@ -1,0 +1,11 @@
+package utils
+
+import (
+	"log"
+)
+
+func CheckErr(err error) {
+	if err != nil {
+		log.Fatal(err)
+	}
+}
